pyproject.toml
[project]
name = "colorize"
version = "0.1.0"
description = "Diffusion-based image colorization with super-resolution and model compression"
requires-python = ">=3.12"
dependencies = [
    "torch>=2.2.0",
    "torchvision>=0.17.0",
    "numpy",
    "opencv-python",
    "Pillow",
    "datasets",
    "wandb",
    "lpips",
    "scikit-image",
    "tqdm",
    "accelerate",
    "fastapi",
    "uvicorn",
    "python-multipart",
    "gradio",
    "onnx",
    "onnxruntime",
    "onnxconverter-common",
]
